requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
sqlalchemy==1.4.23
torch==1.9.0
torchvision==0.10.0
mlflow==1.20.2
reportlab==3.6.1
Pillow==8.3.2